+++ api-ud/src/main/java/com/web/app/mapper/UD08HomologationVariablesMapper.java
@@ -20,65 +20,65 @@ import java.util.List;
 @Mapper
 public interface UD08HomologationVariablesMapper {
 
-    /**
-     * 查询所有产品类别
-     *
-     * @return 产品类别列表
-     */
-    List<ProductClassMaster> selectAllProductClass();
+        /**
+         * 查询所有产品类别
+         *
+         * @return 产品类别列表
+         */
+        List<ProductClassMaster> selectAllProductClass();
 
-    /**
-     * 查询所有市场
-     *
-     * @return 市场列表
-     */
-    List<MarketMaster> selectAllMarket();
+        /**
+         * 查询所有市场
+         *
+         * @return 市场列表
+         */
+        List<MarketMaster> selectAllMarket();
 
-    /**
-     * 检查HDOC变量是否存在
-     *
-     * @param variables 变量名
-     * @return 记录数
-     */
-    Integer countHdocVariables(@Param("variables") String variables);
+        /**
+         * 检查HDOC变量是否存在
+         *
+         * @param variables 变量名
+         * @return 记录数
+         */
+        Integer countHdocVariables(@Param("variables") String variables);
 
-    /**
-     * 根据PC+NUM+MARKET查询用户定义规则数量
-     *
-     * @param pc     PC代码
-     * @param num    序号
-     * @param market 市场
-     * @return 记录数
-     */
-    Integer countUserDefinedRule(@Param("pc") String pc,
-            @Param("num") Integer num,
-            @Param("market") String market);
+        /**
+         * 根据PC+NUM+MARKET查询用户定义规则数量
+         *
+         * @param pc     PC代码
+         * @param num    序号
+         * @param market 市场
+         * @return 记录数
+         */
+        Integer countUserDefinedRule(@Param("pc") String pc,
+                        @Param("num") Long num,
+                        @Param("market") String market);
 
-    /**
-     * 插入用户定义规则
-     *
-     * @param rule 规则实体
-     * @return 影响行数
-     */
-    Integer insertUserDefinedRule(HdocUserDefinedRules rule);
+        /**
+         * 插入用户定义规则
+         *
+         * @param rule 规则实体
+         * @return 影响行数
+         */
+        Integer insertUserDefinedRule(HdocUserDefinedRules rule);
 
-    /**
-     * 更新用户定义规则
-     *
-     * @param rule 规则实体
-     * @return 影响行数
-     */
-    Integer updateUserDefinedRule(HdocUserDefinedRules rule);
+        /**
+         * 更新用户定义规则
+         *
+         * @param rule 规则实体
+         * @return 影响行数
+         */
+        Integer updateUserDefinedRule(HdocUserDefinedRules rule);
 
-    /**
-     * 删除用户定义规则
-     *
-     * @param pc     PC代码
-     * @param num    序号
-     * @param market 市场
-     * @return 影响行数
-     */
-    Integer deleteUserDefinedRule(@Param("pc") String pc,
-            @Param("num") Integer num,
-            @Param("market") String market);
+        /**
+         * 删除用户定义规则
+         *
+         * @param pc     PC代码
+         * @param num    序号
+         * @param market 市场
+         * @return 影响行数
+         */
+        Integer deleteUserDefinedRule(@Param("pc") String pc,
+                        @Param("num") Long num,
+                        @Param("market") String market);
 }
